add subsystem vehicle_counter
