+++ run.sh
@@ -1,4 +1,4 @@
 # export PATH=/media/trung/01D4B61EC6C65DB0/miniconda3/bin:$PATH
 # python -W ignore vehicle_detection_main.py
-python -W ignore gui.py
 # tensorboard --logdir="graphs"
+python -W ignore gui.py
