+++ traffic_measurement.csv
@@ -1,8 +1,6 @@
 cam_id,date,start_time,end_time,num_vehicles
-cam1,10/12/2019,1,2,3
-cam2,10/12/2019,1,2,8
-cam3,11/12/2019,1,2,6
-cam3,10/12/2019,1,2,6
-cam1, 10/12/2019, 1, 2, 3
-cam2, 10/12/2019, 1, 2, 8
-cam2, 10/12/2019, 1, 2, 8
+cam2,11/12/2019,23,24,8
+cam3,11/12/2019,23,24,15
+cam1,11/12/2019,23,24,3
+cam2,12/12/2019,12,13,8
+cam2,12/12/2019,14,15,8
