+++ vehicle_counter/detector/__init__.py
@@ -1,0 +1,8 @@
+from abc import ABC, abstractmethod
+
+##Interface
+class Detector(ABC):
+    
+    @abstractmethod
+    def get_detections(self, image):
+        pass
